Move `vecmem::memory_order` into its own header
This commit moves `vecmem::memory_order` into its own header in order to
avoid circular dependencies in #275 and #276.
