+++ core/include/vecmem/memory/impl/device_atomic_ref.ipp
@@ -7,6 +7,9 @@
  */
 #pragma once
 
+// vecmem includes
+#include "vecmem/memory/memory_order.hpp"
+
 // HIP include
 #if defined(__HIP_DEVICE_COMPILE__)
 #include <hip/hip_runtime.h>
